pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dash-dynamic-images"
version = "1.0.0"
authors = [
    { name="Tomasz Rewak", email="tomasz.rewak@gmail.com" },
]
description = "Enables loading react components in Dash applications directly from local project files, without any need for a separate build process."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'dash >= 2.6.0, < 3',
    'Pillow >= 9.5.0, < 10',
]

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
"Homepage" = "https://github.com/TomaszRewak/dash-dynamic-images"
"Bug Tracker" = "https://github.com/TomaszRewak/dash-dynamic-images/issues"

[tool.mypy]
python_version = "3.10"
warn_return_any = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]